requirements.txt
asgiref
certifi
cffi
charset-normalizer
click
coreapi
coreschema
cryptography
defusedxml
Django
django-cors-headers
django-environ
django-templated-mail
djangorestframework
environ
gunicorn
h11
idna
itypes
Jinja2
MarkupSafe
oauthlib
psycopg2-binary
pycparser
python3-openid
pytz
requests
requests-oauthlib
six
social-auth-app-django
social-auth-core
sqlparse
uritemplate
urllib3